Refactor and create router for task model
