+++ src/index.js
@@ -3,142 +3,16 @@ require("./db/mongoose");
 const User = require("./models/user");
 const Task = require("./models/task");
 
+const userRouter = require('./routers/user-routers')
+const tasRouter = require('./routers/task-router')
+
 const app = express();
 const port = process.env.port || 3000;
 
 app.use(express.json());
+app.use(userRouter)
+app.use(tasRouter)
 
-app.post("/users", async (req, res) => {
-  const user = new User(req.body);
-  try {
-    await user.save();
-    res.status(201).send(user);
-  } catch (e) {
-    res.status(400).send(e);
-  }
-});
-
-app.post("/tasks", async (req, res) => {
-  const task = new Task(req.body);
-
-  try {
-    await task.save();
-    res.status(201).send();
-  } catch (e) {
-    res.status(400).send(e);
-  }
-});
-
-app.get("/users", async (req, res) => {
-  try {
-    const users = await User.find({});
-    res.send(users);
-  } catch (e) {
-    res.status(500).send();
-  }
-});
-
-app.get("/users/:id", async (req, res) => {
-  const _id = req.params.id;
-
-  try {
-    const user = await User.findById(_id);
-    if (!user) {
-      return res.status(404).send();
-    }
-    res.send(user);
-  } catch (e) {
-    res.status(500).send();
-  }
-});
-
-app.get("/tasks", async (req, res) => {
-  try {
-    const tasks = await Task.find({})
-    res.status(201).send(tasks)
-  } catch (e) {
-    res.status(500).send()
-  }
-})
-
-app.get("/tasks/:id", async (req, res) => {
-  const _id = req.params.id;
-
-  try {
-    const task = await Task.findById(_id)
-    if (!task) {
-      return res.status(404).send()
-    }
-    res.send(task)
-  } catch (e) {
-    res.status(500).send()
-  }
-})
-
-app.patch('/users/:id', async (req, res) => {
-  const updates = Object.keys(req.body)
-  const allowedUpdates = ["name", "email", "password", "age"]
-  const isValidOperation = updates.every((update) => allowedUpdates.includes(update))
-
-  if (!isValidOperation) {
-    return res.status(400).send({ 'error': 'Invalid update!' })
-  }
-  try {
-    const user = await User.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true })
-    if (!user) {
-      res.status(404).send()
-    }
-    res.send(user)
-  } catch (e) {
-    res.status(400).send(e)
-  }
-})
-
-app.patch('/tasks/:id', async (req, res) => {
-  const updates = Object.keys(req.body)
-  const allowedUpdates = ["description", "completed"]
-  const isValidOperation = updates.every((update) => allowedUpdates.includes(update))
-
-  if (!isValidOperation) {
-    return res.status(400).send({ 'error': 'Invalid update operation !' })
-  }
-
-  try {
-    const task = await Task.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true })
-
-    if (!task) {
-      res.status(404).send()
-    }
-    res.send(task)
-  } catch (e) {
-    res.status(500).send(e)
-  }
-})
-
-app.delete('/users/:id', async (req, res) => {
-
-  try {
-    const user = await User.findByIdAndDelete(req.params.id)
-    if (!user) {
-      res.status(404).send()
-    }
-    res.send(user)
-  } catch (e) {
-    res.status(500).send()
-  }
-})
-
-app.delete('/tasks/:id', async (req, res) => {
-  try {
-    const task = await Task.findByIdAndDelete(req.params.id)
-    if (!task) {
-      res.status(404).send()
-    }
-    res.send(task)
-  } catch (e) {
-    res.status(500).send()
-  }
-})
 
 app.listen(port, () => {
   console.log("Server is listening on port " + port);
